separate tooltip parsing test from other parsing tests
